bcrypto: more compat. more hash id improvements.

+++ src/pbkdf2/pbkdf2.h
@@ -13,12 +13,12 @@ bool
 bcrypto_pbkdf2(
   const char *name,
   const uint8_t *data,
-  uint32_t datalen,
+  size_t datalen,
   const uint8_t *salt,
-  uint32_t saltlen,
+  size_t saltlen,
   uint32_t iter,
   uint8_t *key,
-  uint32_t keylen
+  size_t keylen
 );
 
 #if defined(__cplusplus)
